refactor: AI プロンプト文字列を専用ファイルに抽出
- 長いプロンプト文字列を ai-prompts.ts に移動
- createTaskEstimationPrompt() 関数を追加
- createScheduleRecommendationPrompt() 関数を追加
- SYSTEM_PROMPTS 定数を追加
- コードの保守性とテスタビリティを向上
